Update deleteProduct Controller to handel invalid product id

+++ backend/controllers/product.controller.js
@@ -76,6 +76,12 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
   // console.log(id);
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    console.log("Invalid Product id!");
+    return res
+      .status(404)
+      .json({ success: false, message: "Invalid Product id!" });
+  }
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({
@@ -84,6 +90,6 @@ export const deleteProduct = async (req, res) => {
     });
   } catch (error) {
     console.error("Error deleting product:", error.message);
-    res.status(404).json({ success: false, message: "Product Not Found!" });
+    res.status(500).json({ success: false, message: "Product Not Found!" });
   }
 }
